Break out getClaimStatus into separate file.
