+++ utils/getScenarioContent.tsx
@@ -7,8 +7,9 @@
  * description in ScenarioTypeNames for easy(ish) reference.
  */
 
-import { Claim, ClaimDetailsContent, ClaimStatusContent, I18nString, ScenarioContent } from '../types/common'
+import { Claim, ClaimDetailsContent, ScenarioContent } from '../types/common'
 import getClaimDetails from './getClaimDetails'
+import getClaimStatus from './getClaimStatus'
 
 export enum ScenarioType {
   Scenario1,
@@ -62,28 +63,6 @@ export function getScenario(claimData: Claim): ScenarioType {
 }
 
 /**
- * Get Claim Status description content.
- */
-export function getClaimStatusDescription(scenarioType: ScenarioType): I18nString {
-  return `claim-status:scenarios.${ScenarioType[scenarioType].toLowerCase()}.description`
-}
-
-/**
- * Get Claim Status content.
- */
-export function buildClaimStatusContent(scenarioType: ScenarioType): ClaimStatusContent {
-  const statusContent: ClaimStatusContent = {
-    statusDescription: getClaimStatusDescription(scenarioType),
-    nextSteps: [
-      'Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
-      'Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.',
-    ],
-  }
-
-  return statusContent
-}
-
-/**
  * Return scenario content.
  */
 export default function getScenarioContent(claimData: Claim): ScenarioContent {
@@ -91,7 +70,7 @@ export default function getScenarioContent(claimData: Claim): ScenarioContent {
   const scenarioType = getScenario(claimData)
 
   // Construct claim status content.
-  const statusContent = buildClaimStatusContent(scenarioType)
+  const statusContent = getClaimStatus(scenarioType)
 
   // Construct claim details content.
   if (!claimData.claimDetails) {
